add hw4

+++ service/contacts.js
@@ -1,19 +1,25 @@
-const Contact = require("../model/contact_schema");
+const Contact = require("../model/schemas/contact_schema");
 
-const getListContacts = async () => {
-  const contacts = await Contact.find({});
+const getContactsList = async (userId) => {
+  const contacts = await Contact.find({owner: userId }).populate({
+    path: 'owner',
+    select:'name email phone -_id'
+  });
 
   return contacts;
 };
 
-const getContactById = async (id) => {
-  const contact = await Contact.findOne({ _id: id });
+const getContactById = async (id, userId) => {
+  const contact = await Contact.findOne({ _id: id, owner: userId }).populate({
+    path: 'owner',
+    select:'name email phone -_id',
+  });
 
   return contact;
 };
 
-const removeContact = async (id) => {
-  const contact = await Contact.findByIdAndRemove({ _id: id });
+const removeContact = async (id, userId) => {
+  const contact = await Contact.findByIdAndRemove({_id:id, owner: userId });
 
   return contact;
 };
@@ -24,29 +30,21 @@ const addContact = async (body) => {
   return contact;
 };
 
-const updateContact = async (id, body) => {
+const updateContact = async (id, body, userId) => {
   const contact = await Contact.findByIdAndUpdate(
-    { _id: id },
+    {_id: id, owner: userId },
     { ...body },
     { new: true }
   );
 
   return contact;
 };
-const updateStatusContact = async (id, body) => {
-  const contact = await Contacts.findByIdAndUpdate(
-    { _id: id },
-    { ...body },
-    { new: true },
-  );
 
-  return contact;
-};
 module.exports = {
-  getListContacts,
+  getContactsList,
   getContactById,
   removeContact,
   addContact,
   updateContact,
-  updateStatusContact
+  
 };
